requirements.txt
cx_Oracle
pandas
scikit-learn
numpy
openai
python-dotenv
openpyxl
joblib
